data_structures::counter

Struct Counter

pub struct Counter<T>
where
    T: Eq + Hash,
{ /* private fields */ }

Counter

Underlying structure is a hashmap, it increments the value by 1 for every occurrence of the key.

Implementations

impl<T: Eq + Hash> Counter<T>

pub fn new() -> Self

pub fn len(&self) -> usize

Checks the length of a Counter

Returns the number of keys in the Counter
Return type is usize

Example

use data_structures::Counter;
let mut counter: Counter<&str> = Counter::new();
counter.add("A");
counter.add("B");
counter.add("C");
counter.add("A");

assert_eq!(counter.len(), 3);

pub fn is_empty(&self) -> bool

Checks if a Counter has any keys

Checks if the Counter is empty Returns a boolean

Example

use data_structures::Counter;
let mut counter: Counter<&str> = Counter::new();

assert!(counter.is_empty());

pub fn add(&mut self, item: T)

Adds a key or increments count of present key

If a key is already present, then add increments the count by 1.
If a key is not present, then add inserts the key and sets the count to 1.

Example

use data_structures::Counter;
let mut counter: Counter<&str> = Counter::new();

counter.add("A");
assert_eq!(counter.get("A"), Some(&1));

counter.add("A");
assert_eq!(counter.get("A"), Some(&2));

pub fn subtract(&mut self, item: T)

Subtracts the count of a key

Decrease the count of a key by 1. If the count reaches zero, then the key is removed entirely.

Nothing occurs if the key is not present.

Examples:

use data_structures::Counter;
let mut counter: Counter<char> = "rustisgreat".chars().collect();

assert_eq!(counter.get(&'t'), Some(&2));
counter.subtract('t');
assert_eq!(counter.get(&'t'), Some(&1));
assert_eq!(counter.get(&'d'), None);

pub fn get<Q>(&self, key: &Q) -> Option<&usize>
where T: Borrow<Q>, Q: Hash + Eq + ?Sized,

Gets the count of a key

Returns the count of the key as an Option<&usize>

Example:

use data_structures::Counter;
let mut counter: Counter<char> = "rust".chars().collect();
let r_count = match counter.get(&'r') {
    Some(&r_count) => r_count,
    None => 0,
};

assert_eq!(r_count, 1)

pub fn subtract_counter(&mut self, other: &Counter<T>)
where T: Clone,

Subtracts the counts of another Counter from this one.

For each key in other, its count is subtracted from the corresponding count in self. Keys whose count reaches zero are removed entirely. Keys present in self but not in other are left unchanged.

Example:

use data_structures::Counter;
let mut counter1: Counter<char> = "rust".chars().collect();
let mut counter2: Counter<char> = "rustacean".chars().collect();

counter2.subtract_counter(&counter1);

assert_eq!(counter2.get(&'r'), None);
assert_eq!(counter2.get(&'u'), None);
assert_eq!(counter2.get(&'s'), None);
assert_eq!(counter2.get(&'t'), None);
assert_eq!(counter2.get(&'a'), Some(&2));
assert_eq!(counter2.get(&'c'), Some(&1));
assert_eq!(counter2.get(&'e'), Some(&1));
assert_eq!(counter2.get(&'n'), Some(&1));

pub fn add_counter(&mut self, other: &Counter<T>)
where T: Clone,

Adds the counts of another Counter to this one

For each key in other, its count is added to teh corresponding count in self. Keys present in other but not in self are inserted with other’s count.

Counts saturate at usize::MAX rather than overflowing.

Example:

use data_structures::Counter;
let mut counter1: Counter<char> = "rust".chars().collect();
let mut counter2: Counter<char> = "rustacean".chars().collect();

counter1.add_counter(&counter2);

assert_eq!(counter1.get(&'r'), Some(&2));
assert_eq!(counter1.get(&'u'), Some(&2));
assert_eq!(counter1.get(&'s'), Some(&2));
assert_eq!(counter1.get(&'t'), Some(&2));
assert_eq!(counter1.get(&'a'), Some(&2));
assert_eq!(counter1.get(&'c'), Some(&1));
assert_eq!(counter1.get(&'e'), Some(&1));
assert_eq!(counter1.get(&'n'), Some(&1));

pub fn remove<Q>(&mut self, key: &Q)
where T: Borrow<Q>, Q: Hash + Eq + ?Sized,

Removes key from Counter entirely

Removes key and its count (regardless of count value) from Counter. If key does not exist, then nothing happens.

Example

use data_structures::Counter;
let mut counter: Counter<char> = Counter::new();
counter.add('a');
counter.add('a');
counter.add('b');
counter.add('c');

counter.remove(&'a');
assert_eq!(counter.get(&'a'), None);
assert_eq!(counter.get(&'b'), Some(&1));
assert_eq!(counter.get(&'c'), Some(&1));

pub fn keys(&self) -> impl Iterator<Item = &T>

Returns an iterator over the keys of the Counter.

Example:

use data_structures::Counter;
let counter: Counter<char> = "rust".chars().collect();

for count in counter.keys() {
    println!("{}", count);
}

pub fn values(&self) -> impl Iterator<Item = &usize>

Returns an iterator over the counts of the Counter

Example

use data_structures::Counter;
let counter: Counter<char> = "rust".chars().collect();

for count in counter.values() {
    println!("{}", count);
}

pub fn counts(&self) -> impl Iterator<Item = &usize>

Alias for Counter::values

pub fn most_common(&self) -> Option<&T>

Finds the key with the highest count

Returns the name of the key with the highest count. If multiple keys have the same count, it takes the first one encountered.

Please note: “first encountered” can be different every time since a Counter is unordered.

Returns None if Counter is empty

Example:

use data_structures::Counter;
let counter1: Counter<char> = "rustacean".chars().collect();
let counter2: Counter<char> = Counter::new();

assert_eq!(counter1.most_common(), Some(&'a'));
assert_eq!(counter2.most_common(), None);

pub fn least_common(&self) -> Option<&T>

Finds the key with the lowest count

Returns the name of the key with the lowest count. If multiple keys have the same count, it takes the first one encountered.

Please note: “first encountered” can be different every time since a Counter is unordered.

Returns None if Counter is empty

Example:

use data_structures::Counter;
let mut counter: Counter<char> = "abbccc".chars().collect();

assert_eq!(counter.least_common(), Some(&'a'));

pub fn highest_count(&self) -> Option<usize>

Finds the highest count in the Counter

If there are multiple counts that are the highest, it returns the first encountered.

Please note: “first encountered” can be different every time since a Counter is unordered.

Example:

use data_structures::Counter;
let mut counter: Counter<char> = "rustacean".chars().collect();

assert_eq!(counter.highest_count(), Some(2));

pub fn lowest_count(&self) -> Option<usize>

Finds the lowest count in a Counter

If there are multiple counts that are the lowest, it returns the first encountered.

Please note: “first encountered” can be different every time since a Counter is unordered.

Example:

use data_structures::Counter;
let mut counter: Counter<&str> = Counter::new();
counter.add("A");
counter.add("A");
counter.add("B");
counter.add("B");
counter.add("C");

assert_eq!(counter.lowest_count(), Some(1));

pub fn total_count(&self) -> usize

Returns the sum of all counts in the Counter.

Returns 0 if the Counter is empty.

Example:

use data_structures::Counter;
let counter: Counter<char> = "rustacean".chars().collect();

assert_eq!(counter.total_count(), 9);